:art: Apply active profiles consistently with AOT
Profiles that are active during AOT processing are automatically enabled
when the AOT initializer runs. While this works for an arrangement that
only relies on the ApplicationContext, it does not for Spring Boot that
has specific handling of profiles when it prepares the environment, way
before the ApplicationContext is event created.

This commit adds a specific contribution that generates a dedicated
EnvironmentPostProcessor. It also updates the handling of post
processors so that when AOT runs, the AOT generated one if it exists is
invoked first.

This has the effect of consistently activating such profiles in a Infra application.