:art: Ensure descendants are always recalculated on cache refresh
Fix `DefaultIterableConfigurationPropertySource` to ensure that cached
descendants are always updated on a refresh.

The cache implementation assumes that it is safe to reuse previous
`mappings` and `reverseMappings` data since it doesn't matter if
superfluous values are included. For `descendants` however, we always
want to recalculate values so that we don't get false positives.

Unfortunately, prior to this commit, we only updated the descendants
if a reverseMapping was added. The meant that on a cache refresh,
existing descendants were removed.